+++ db/migrate/20250918090000_add_primary_strategy_key_to_tournaments.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class AddPrimaryStrategyKeyToTournaments < ActiveRecord::Migration[7.2]
+  def change
+    add_column :tournaments, :primary_strategy_key, :string, null: false, default: 'points'
+  end
+end
+
+